@@ -1,7 +1,7 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
 import PropTypes from 'prop-types';
-import { withStyles } from '@material-ui/core/styles';
+import { withStyles } from '@material-ui/styles';
 import Grid from '@material-ui/core/Grid';
 import Paper from '@material-ui/core/Paper';
 import Typography from '@material-ui/core/Typography';
@@ -23,7 +23,7 @@ const styles = theme => ({
         textDecoration: 'none',
     },
     root: {
-        padding: `${theme.spacing.unit * 2}px ${theme.spacing.unit * 2}px`,
+        padding: `${theme.spacing(2)}px ${theme.spacing(2)}px`,
         transition: theme.transitions.create(['background-color'], {
             duration: theme.transitions.duration.standard,
         }),
@@ -43,14 +43,14 @@ const styles = theme => ({
         color: textGreyColor,
         width: 12,
         height: 12,
-        marginLeft: theme.spacing.unit,
+        marginLeft: theme.spacing(1),
     },
 });
 
 const RepositoryCard = ({ classes, repository }) => (
     <Link to={`/${repository.full_name}/${repository.default_branch}`} className={classes.link}>
         <Paper className={classes.root} elevation={1}>
-            <Grid container style={{ flex: 1 }} spacing={8} alignItems="center">
+            <Grid container style={{ flex: 1 }} spacing={1} alignItems="center">
                 <Grid item>
                     <UserAvatar username={repository.owner.login} />
                 </Grid>
